Factory Girl triumphs where FixtureReplacement failed
Remove the comment about Factory Girl behaving the same as
FixtureReplacement does; it actually handles this case perfectly.

This is because when defining a factory, the variable passed into the
block is not an ActiveRecord instance but actually a Factory class
instance, effectively a proxy.

So, while the block defines:

  u.passphrase { Sham.passphrase }
  u.passphrase_confirmation { Sham.passphrase }

These methods are not actually invoked on the ActiveRecord instance.
Rather, when we call either User.make or User.make! with overrides,
those overrides are applied to the attributes instead of the default
values.

That is, while FixtureReplacement with overrides effectively did:

  user.passphrase = default
  user.passphrase = override

And the override had no effect due to the logic in the model's setter
method, Factory Girl does:

  user.passphrase = override

And gives us the expected/desired result.

Signed-off-by: Wincent Colaiuta <win@wincent.com>