Remove Bundler warning from spec_helper.rb
Commit c4aa8284 goes into great detail explaining the differences
between running:

  rspec ...

And:

  bundle exec rspec ...

Given that the load path is set up in exactly the same way in both
cases, it seems harmless enough to allow the user to just do "rspec ..."
without pestering them with a warning. To all intents and purposes the
environment should work basically identically, especially now that we've
worked around our RUBYOPT bug which previously only affected the "bundle
exec ..." and "rake spec ..." cases.

If the user ends up running the wrong version of RSpec then they'll be
alterted anyway because Bundler will raise:

    You have already activated rspec-core 2.0.0.beta.13, but your Gemfile
    requires rspec-core 2.0.0.beta.14. Consider using bundle exec.
    (Gem::LoadError)

So, while it is a good habit to get in to to always use the Rake tasks
or "bundle exec", the warning is probably unnecessarily scary, so just
remove it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>