Remove Reset factory kludge
Setting the email_address attribute was never really needed to produce a
valid record using the factory.

Reset.make produces an unsaved record, so the validation in question
does not run.

Reset.make! produces a saved record, but as this is the first time it
gets saved the validation doesn't run here either (it only runs on
updates of existing, saved records).

Signed-off-by: Wincent Colaiuta <win@wincent.com>