Track Git repo of rspec-mocks
We need a fix that isn't yet in a released version of the Gem.

Even though we are using RR and don't directly require "rspec/mocks" at
any time, rspec-rails itself depends on rspec, and rspec depends on
rspec-mocks, so Bundler ends up requiring it for us.

Even if this weren't the case, Steak has a "require 'rspec'" in it and
that ends up doing a "require 'rspec/mocks'".

This effectively pollutes the global namespace with methods from the
RSpec mocking library (like should_receive and should_not_receive), and
also adds methods which clash with methods with the same name in the RR
library (like stub and mock).

So we need the fix from upstream which effectively turns "require
'rspec/mocks'" into a "no-op", because it doesn't inject anything into
the global namespace until the user explicitly requests it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>