Use loser param expectations in topic mocks
Put a very lose restriction on the range of accepted parameters, because
when we stub a method with "@topic" as the parameter and then call
"@topic.save" it is possible that we end up altering the topic in a way
that it is no longer considered "equal" to the one we passed in to the
message expectation.

This was part of an attempt to fix the following bug, but it turns out
that there is weirdness in the interaction between RR and ActionMailer,
and it is _never_ possible to set a return value or to evaluate a block.
Hopefully a solution will be forthcoming:

See:

http://github.com/btakita/rr/issues/#issue/34

Signed-off-by: Wincent Colaiuta <win@wincent.com>