Add map_from matcher and use it in routing specs
This is the inverse operation of the map_to matcher.

I like the idea of explicitly showing that we are doing a two-way test:

  get('/wiki').should   map_to({ :controller => 'articles', :action => 'index' })
  get('/wiki').should map_from({ :controller => 'articles', :action => 'index' })

Rather than the official, implicit method:

  { :get => '/wiki' }.should route_to({ :controller => 'articles', :action => 'index' })

I may still add a "map_both" matcher or similar to do this, but I am
still undecided about it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>