Flesh out article routing specs
Here we continue using the new "API" to see if it is more expressive and
pleasant to use.

So far, so good. It is tempting to avoid some of the repetition:

  get('/wiki').should   map_to({ :controller => 'articles', :action => 'index' })
  get('/wiki').should map_from({ :controller => 'articles', :action => 'index' })
  articles_path.should == '/wiki'

Could be rewritten as:

  path   = '/wiki'
  params = { :controller => 'articles', :action => 'index' }
  get(path).should   map_to(params)
  get(path).should map_from(params)
  articles_path.should == path

But while this saves us from repeating some literal strings and hashes,
it is much less "scannable".

Likewise, I could add a "map_both" matcher, and probably will, but if I
do that then I suspect that the "map_from" matcher will basically never
be used.

Signed-off-by: Wincent Colaiuta <win@wincent.com>