Add "map" matcher and use it in router specs
I am not convinced that this is the right name, but given that RESTful
routes are so common, we should optimize for that case and just use
"map" rather than repeating identical "map_to" and "map_from"
assertions.

Alternative names for the matcher include "map_as", "map_via" and
"map_both".

Signed-off-by: Wincent Colaiuta <win@wincent.com>