Move url_for_redirect into the model
This makes it more easily unit testable, and will allow us to share
regular expressions with the validations.

As proof that it will be easier to provide consistent behavior, note how
that in moving it I fixed a discrepancy (relative paths were allowed in
the model but not in the controller).

The only downside is that in the model we don't have access to routing
helpers, so must hardcode "/wiki" rather than using articles_path.

Signed-off-by: Wincent Colaiuta <win@wincent.com>