deploy: use explicit "submodule update"
Seeing as we may end up checking out arbitrary revisions, not just the
current HEAD, we'll need to do a "submodule update" anyway to be sure
that we have the right submodule version checked out.

So this means we may as well drop the "--recursive" flag to "git clone"
which was recently added.

Signed-off-by: Wincent Colaiuta <win@wincent.com>