Triage JsController specs
Unfortunately there is a bug in Rails which prevents us from doing any
JsController testing at all; it reports "No route matches" errors
despite the fact that the controller works in the browser, and the
separate routing specs also pass.

The likely cause is an inability of the Rails functional test machinery
to target actions that are mapped using the "match" operator. This seems
to be a new bug to Rails 3, either the new routing DSL or something else
in the framework, as in previous versions of Rail there was no problem
at all in routing to such actions.

Signed-off-by: Wincent Colaiuta <win@wincent.com>