Unbreak tweets#index view specs (HTML format)
Can't say I am happy with the stubbing that is necessary in the "before"
block in order to get this working:

  stub(view).render('tweets.html.haml')
  stub.proxy(view).render({:template => 'tweets/index.html.haml'}, {})

First we must stub the call to render the partial, because we test that
separately in another file, we don't want errors in that partial
bubbling up into this view's specs, and we don't want to have to do the
additional set-up required by the partial.

But then we have to stub and proxy the call to "render" of the template
being tested itself, which is ugly due to the format of the parameters.
I would like to find a better way.

Signed-off-by: Wincent Colaiuta <win@wincent.com>