Partially unbreak topics#show view specs
As noted in the commit body, helper methods injected using
"helper_method" in the controller are not included in the view, so this
requires us to do some problematic stubbing. Hopefully a fix will
become available; I've tried to find a solution but have been
unsuccessful so far.

On the bright side, notice how in RSpec 2 we can drop the explicit
"assign" calls during set-up.

Signed-off-by: Wincent Colaiuta <win@wincent.com>