Add real sortable_header_cell calls to issues/_issues spec
Commit eb700bc fixed views which mistakenly used:

  -sortable_header_cell

Rather than:

  =sortable_header_cell

In order to test this we need to get rid of our stubs and do the
actual testing of the rendered output.

I really need to do the same for all such views that were broken, but
the tests are quite repetitive. I am not sure whether it is worth the
time investment required.

Signed-off-by: Wincent Colaiuta <win@wincent.com>