Move workaround into controller to minimize impact on specs
A few specs pass in an array here for ease of test setup, and this will
break if we try sending it an "all" message. So the easiest fix, then,
is just to move the "all" call up into the controller.

Ideally the ActiveRecord::Relations don't effect their queries until
actually used (ie in the view), but as this is a temporary workaround
for a bug in Rails that will hopefully be fixed, it doesn't really
matter.

Signed-off-by: Wincent Colaiuta <win@wincent.com>