Add Tag#find_with_tag_names method
This will enable us to eliminate some repetition. Note that although it
is done using the new query API, we call "to_a" at the end to maintain
the existing semantics of "find" methods (which return real arrays).

I am not 100% sure that this is the right thing to do, but will leave it
as-is for now. The alternative is to change this spec:

  Tag.find_with_tag_names('foo', 'bar', 'baz').should =~ [@foo, @bar, @baz]

To either sort the results first or call "to_a".

Signed-off-by: Wincent Colaiuta <win@wincent.com>