Remove email deletion handling from users#edit
It's not active in the view, so not much sense in including it here
either.

Side benefit: no more race conditions to worry about.

Seeing as we never really deleted emails, only marked them as "deleted",
seems silly to offer it anyway. If it's requested, can look at doing it
in another way (probably via the emails controller). In the meantime,
YAGNI.

This means we can do some further clean up too, such as getting rid of
the deleted_at column in the emails table, and making some of our
queries simpler.

Signed-off-by: Wincent Colaiuta <win@wincent.com>