Restore syntax highlighting functionality
This was broken during refactoring. The problem is that we add the
"highlighted" class to guard against double-highlighting, and that turns
our class string into "ruby-syntax highlighted" instead of the expected
"ruby-syntax".

While this works, a more robust solution will be to ignore all other
classes and specifically look for a single class of the form
"lang-syntax".

See:

  https://wincent.com/issues/1605

Signed-off-by: Wincent Colaiuta <win@wincent.com>