"bundle update" to pull down RSpec 2.0.0.beta.20
Note that webrat is no longer implicitly pulled down (due to the RSpec
dependency that existed up to beta 19), so we have to specify it
explicitly in order to keep our specs passing.

In the future may investigate if/how Capybara matchers could be used in
those view specs which currently use Webrat matchers, so as to drop one
more dependency.

Signed-off-by: Wincent Colaiuta <win@wincent.com>