Add Unicorn to bundle to avoid Rack clashes
Unicorn loads before Rails, and without Bundler, so it pulls in the most
appropriate version of Rack on the system; in the case of the current
staging environment, that means Rack 1.0.1.

But Rails 3 uses Bundler, and it requires and includes Rack 1.2.5. So
when Unicorn tries to load the Rails app, an exception is thrown on
trying to activate Rack 1.2.5 because Rack 1.0.1 is already activated.
The solution then, is to include Unicorn itself in the bundle.

Signed-off-by: Wincent Colaiuta <win@wincent.com>