Add memcache gem to Gemfile
Turns out that my existing sidebar caching is completely broken; it
isn't hitting memcached at all (ie. I can start the server up in the
production environment to ensure that caching will actually happen, and
then I can add new records and observer how everything works perfectly
even if I kill the memcached server! It's simply not being accessed at
all.)

Noticed while trying to get unicorn to launch on staging server; its
config file in the after_fork hook accesses
ActiveSupport::Cache::MemCacheStore, which causes Rails to try requiring
the memcache gem, and that fails.

Will need to configure Rails to use memcached, but only for fragment
caching, and unfortunately that is not officially documented anywhere
that I can see, but there is unofficial and apparently out-of-date
documentation out there like this page:

  http://ariekanarie.nl/archives/46/using-memcached-to-speed-up-fragment-caching

Which says you have to set it up manually in this way:

  memcache_options = {
    :c_threshold = 10_00,
    :compression = true,
    :debug = false,
    :namespace = 'yourappname_or_anything_you_like',
    :readonly = false,
    :urlencode = false
  }
  CACHE = MemCache.new(memcache_options)
  CACHE.servers = '127.0.0.1:11211'
  config.action_controller.fragment_cache_store = CACHE, {}

But I will leave that for another commit, seeing as it is a can of worms
(grepping the current Rails source shows that this fragment_cache_store
API no longer exists). It may be that page caching will write to the
filesystem even if memcache is set up in the config (at least, this is
what a two-year-old presentation I found says).

Signed-off-by: Wincent Colaiuta <win@wincent.com>