Add specs for nil params to User.digest method
This is a safety net. Nil values should never get into the database
due to validations, _but_ if they do, we not longer have the protection
of NOT NULL contraints in the database.

Although User.digest should barf if either parameter is nil, add these
explicit checks as a safety net. Don't want people authenticating with
blank passphrases just because of a bug in the application.

Signed-off-by: Wincent Colaiuta <win@wincent.com>