Ensure outgoing mail is sent with content-type 'text/plain'
Outgoing mail was all being sent with content-type 'text/html', which
was ruining the display in the receiver's MUA.

Looks like this one snuck through in the move to the new Action Mailer.

See:

  https://wincent.com/issues/1674

Note that only one example was added to each spec file, but in adding
it I changed the nesting and so nearly the whole file is touched in some
way (ie. it looks worse than it is).

Signed-off-by: Wincent Colaiuta <win@wincent.com>