script/deploy: check that db/schema.rb is up-to-date
This is not an infallible test but it is certainly better than nothing:
it runs "rake db:schema:dump" and compares the result to the
"db/schema.rb" file that is checked in to the SCM.

Note that this means it is comparing the current state of the
development database, as it is not really possible for it to get a
schema dump for any other (non-current) state.

This means you can do pathological stuff, but it should be fine for the
normal development case (ie. where your HEAD is the tip of the master
branch, your migrations in the development environment are up to date,
you keep your test database in sync as well, and you assiduously check
in your db/schema.rb file whenever it changes). It is still possible to
think of cases where this test won't really help, but it's still better
than not having any test at all.

  Scenario: forget to run migrations

    - how are your specs passing?
    - and if they're not passing, why are you deploying?

  Scenario: migrate and forget to check in new db/schema.rb

    - test will pick up that current schema doesn't match last schema
      checked in

  Scenario: on feature branch with migrations, and deploying master

    - test will notice that current schema doesn't match master's
      db/schema.rb; this warning may be spurious, but even if it is
      there is no harm in prompting the user to investigate

Signed-off-by: Wincent Colaiuta <win@wincent.com>