Fix non-left-aligned checkboxes in forms
This appears in Chrome on Windows, and Firefox on the Mac, possibly
elsewhere too. The "width" styling on form inputs actually forces the
checkbox to appear centered on some browsers, but its actually just that
it tries to be wide, and the checkbox ends up sitting on the right
within the box that defines the super-wide checkbox.

The solution is to apply "width" styling only to specific elements
(inputs of type 'text' and 'password', and textarea elements).

See:

  https://wincent.com/issues/1672

Signed-off-by: Wincent Colaiuta <win@wincent.com>