Another layer against session fixation attacks
Session fixation attacks are impossible with the architecture that we're
using here (see "git log --reverse -p --grep=fixation" for notes on
that), in the sense that we don't use the Rails session to store any
authentication or authorization data.

Nevertheless, we do still use the Rails session, even if only for
"cosmetic" bits of data such as the "original_uri" to redirect to after
a login, or the flash.

Seeing as the cost of protecting event his worthless little session is
so small, may as well do it.

Note that it's not enough to just call "reset_session"; we have to be
sure that we copy over the values stored in the session afterwards too,
otherwise things like "original_uri" redirects won't work.

See this guide for an excellent write-up of session fixation attacks:

  http://guides.rubyonrails.org/security.html

Signed-off-by: Wincent Colaiuta <win@wincent.com>