Use mem_cache_store in production
Note that we could never share the same memcache server instance among
different environments, even if we wanted to (for example, to perform
testing of the cache in the test environment).

So it is very important to only set it up in the production environment.
If we ever decide to do testing or development with the cache, it must
be "sandboxed" into a completely separate memcached server instance,
listening on a different port.

This commit seeks to fix the bug in which the memcached server isn't
being used at all on the server.

Signed-off-by: Wincent Colaiuta <win@wincent.com>