Fix exception mailer spec breakage
In the test environment there is no real request but instead an instance
of ActionDispatch::TestRequest (or similar, I don't recall for sure
exactly what class it is) with "host" set to "localhost" so that the
"url" method will always return "localhost".

As this is not an end-user-facing URL, it doesn't matter what the host
URL is in the exception notifcations, really, so adjust the spec to
match the actual behavior.

As noted in ticket #1679, this was not failing locally because locally I
have set "host" to "localhost", but it was failing on the server because
there the "host" is set to "wincent.com".

See:

  https://wincent.com/issues/1679

Signed-off-by: Wincent Colaiuta <win@wincent.com>