Don't HTML escape content inside text/plain emails
The problem here is that the Rails template system doesn't "know" what
kind of output we are producing in the mailer (despite the ".text.erb"
extension; not sure if this is a bug), and so it escapes everything that
gets inserted.

Force it not to with the "raw" helper.

Related Rails ticket:

  https://rails.lighthouseapp.com/projects/8994/tickets/4858

Signed-off-by: Wincent Colaiuta <win@wincent.com>