script/deploy: print summary only after enable subcommand is run
Wouldn't want a failed "summary" subcommand to derail a deployment (not
that I am expecting it to fail, but just in case), so put it at the end,
after everything else has run.

This has the added benefit of making the output easier to copy and paste
into the browser, because it will appear almost at the very end (ie.
just before the "lock" output).

Signed-off-by: Wincent Colaiuta <win@wincent.com>