Stop "snowman" from showing up in GET query strings
There is only one place in the app where we submit forms via GET: the
tags#search action.

This is nice because it shows the query in the query string, but Rails
is dirtying it up by inserting the snowman, so we are forced to switch
over to a POST request. (At least the query is very obvious in the view,
so it doesn't matter so much. We can still provide direct links to tag
searches -- ie. https://wincent.com/tags/search?q=foo+bar -- and they
still work.)

The cost, however, is that by switching over to POST, Rails also inserts
request forgery protection tokens. We don't want these here because they
end up contaminating cached pages. And we don't need them either,
because this is a non-destructive request (ie. not a create, update or
destroy).

Seeing as we can't suppress these tokens without turning them off
entirely, we have no choice but to manually build the form by hand,
including inserting the snowman ourselves. I will need to do the same
for the site search form, so will probably add a helper for this.

See:

  https://wincent.com/issues/1685

Signed-off-by: Wincent Colaiuta <win@wincent.com>