Work around acceptance spec failure in server environment
This one described in this ticket:

  https://wincent.com/issues/1666

There is some stale state hanging around between scenarios, but only on
the server. We continue to be "logged in" when starting the new
scenario.

As ensuring a known state (by logging out first) is not a bad thing, do
this. May have to investigate this inter-scenario leakage further down
the track if it is shown to cause problems.

Signed-off-by: Wincent Colaiuta <win@wincent.com>