script/deploy: fix "summary" subcommand with multiple commits
When this was implemented it worked by chance only because the
deployment I tested it on was a one-commit deployment. It fails for
multi-commit deployments.

The solution is to use "git for-each-ref" to obtain a sorted list of
tags, not "git rev-list", which provides a list of all commits reachable
from the tags.

Note that the reverse sort order ensures that annotated tags appear at
the top of the list (non-annotated tags have no taggerdate, so appear at
the other end of the list).

Signed-off-by: Wincent Colaiuta <win@wincent.com>