Eliminate dependency on Hpricot
Seeing as we already pull in Nokogiri (by virtue of our dependence on
Capybara and Webrat), we already have one XML parsing library; we don't
need Hpricot as well.

Nokogiri has a couple of advantages in any case:

   - not abandonware
   - fast
   - actually handles namespaces properly

Note that handling namespaces properly is a double-edged sword, as it
requires us to explicitly name the namespace in all the XPath queries,
but at least it is robust and correct.

Signed-off-by: Wincent Colaiuta <win@wincent.com>