First cut at commit viewing
As it is, it currently works for the standard case but there are some
problems with Git's "--word-diff" output that I fear will make it
unsuitable for doing intra-line colorization.

So this is really just a checkpoint prior to massive refactoring in
which I replace the --word-diff stuff and eventually roll my own.

Signed-off-by: Wincent Colaiuta <win@wincent.com>