Don't let optional attributes trigger db-level constraints
The specs for this were already in place but had a bug. In order to fix
the specs it is not enough to just replace the "nil" parameter with a
blank string value like we would get from a form; we actually have to
save both records to the database (ie. using "make!" instead of just
"make") otherwise the database-level constraint won't kick in.

Signed-off-by: Wincent Colaiuta <win@wincent.com>