Workaround Rails bug #5513, fixing broken constraint on commit :id
This one has already been fixed upstream, after the 3.0.0 release, so
until then the workaround is to add an explicit "nested" block around
the nested resource.

I noticed this because the :commit_id restraint didn't seem to be
working, and any :id was allowed through. The reason why I'd used
:commit_id instead of :id in the first place was that doing the latter
always resulted in "no route found" exceptions.

Rails bug #5513 is the cause, although the problem I'm seeing isn't
explicitly mentioned in that ticket.

I was seeing the nested resource's :id constraint "bubbling up" into the
containing resource. That is, the generated route from:

  resources :repos do
    resources :commit, :id => /foo/
  end

Was, for example:

  repo_commit PUT /repos/:repo_id/commits/:id(.:format)

Mapping to:

  {
    :action => "update",
    :repo_id => /foo/,
    :id => /foo/,
    :controller => "commits"
  }

As you can see, the :id on the nested resources is getting
inappropriately hoisted up into the surrounding context.

Signed-off-by: Wincent Colaiuta <win@wincent.com>

   resources :repos do
-    resources :commits, :commit_id => /[a-f0-9]{7,40}/
+    # Rails BUG:
+    #   https://rails.lighthouseapp.com/projects/8994/tickets/5513
+    # fixed in this commit (post-3.0):
+    #   http://github.com/rails/rails/commit/02480a897be25c24f59180513d37649a31ad3835
+    # until 3.0.1 is released, need to work around this with an explicit
+    # "nested" block:
+    nested do
+      resources :commits, :id => /[a-f0-9]{7,40}/
+    end
   end
   resources :resets
   resources :snippets do