Cacheable flash overhaul (after_filter -> middleware)
As noted in the comments on the new CacheFriendlyFlash middleware,
this commit addresses a long-standing inconsistency: under some
circumstances the "cache_flash" after filter was not running in the
controller specs, even though it does run for real requests.

To make things totally consistent, we get rid of the after filter and
move its functionality into a middleware. In this way our controller
specs can all act as though the middleware didn't exist (as middleware
doesn't run for these fake requests) and our acceptance tests can check
that the middleware actually works.

Signed-off-by: Wincent Colaiuta <win@wincent.com>