Update specs to catch Atom feed problems like issue #1227
These specs falsely passed despite the breakage because we were passing
in an explicit format, thus masking our failure to really set the format
in the controller.

The bug fix not only means that we don't need to do this, but hopefully
the specs will be more resilient to upstream changes as a result (that
is, hopefully the spec will actually fail if a change in Rails breaks
our feeds again).

Signed-off-by: Wincent Colaiuta <win@wincent.com>