Don't run specs outside of staging environment
We are unlikely to make a mistake as utterly stupid as the one GitHub
did here:

  https://github.com/blog/744-today-s-outage

But just in case, be paranoid and only ever run specs and test db
migrations in the staging environment.

This is a stricter separation of "development/test" and "production"
environments than is usually encouraged by the design of Rails (in which
the "environment" is just a set of settings turned no by an environment
variable), and instead is literally a separate environment (in the form
of a completely different machine that can be trashed and thrown away
after staging/testing).

Signed-off-by: Wincent Colaiuta <win@wincent.com>