Fix breakage in controller specs which check "cookies" hash
I'm not sure exactly when this breakage started to occur, but I suspect
it was likely around commit e2600054, which changed out "cache-friendly
flash" mechanism from an after filter into a middleware.

The problem we're seeing here is that the "cookies" hash is always empty
in controller specs, no matter what we do with it or how we read.
Inspection of the response and request objects, however, shows that the
cookie is indeed non-empty, at least according to the
action_dispatch.cookies entry in the environment hash.

I will investigate this further, but in the meantime, get the specs
passing by looking at the environment hash.

Signed-off-by: Wincent Colaiuta <win@wincent.com>