Work around Rails 3.0.3 routing bug
Whoops. Had this workaround for the routing bug introduced in Rails
3.0.3 in my working tree, so all specs were passing, but when deployed
to the server things were horribly broken.

Let this be a lesson to me: it's not enough to just run the spec suite;
you need to run it on the tree that you're actually going to be
deploying. Should probably add a warning to the deployment script about
deploying with a dirty tree.

Signed-off-by: Wincent Colaiuta <win@wincent.com>