Re-order routes for maintainability
The routes file is fairly large now, so organize it basically
alphabetically to make it easier to find things in it.

Note that as part of the clean-up, many instances of "match" were
changed to "get" so as to make the sorting more consistent (ie. so that
all such routes would share the same prefix). It's also shorter which is
nice.

Signed-off-by: Wincent Colaiuta <win@wincent.com>