Remove redundant "about" route
This is already covered by the more general "misc" mapping further down,
so drop it. The "about_path" helper that is generated by this unneeded
mapping was used in only one place in the application anyway, and we
already have precedents for using just literal string mappings for the
misc controller anyway (see the links to "/misc/legal" on the same
layout page).

Signed-off-by: Wincent Colaiuta <win@wincent.com>