Simplify monkey patching of ActiveRecord::Base
Simply re-opening the class and defining our desired method is much
simpler here; the old method is over-engineered and buys us nothing.
After the refactoring, we are still protected against possible changes
to the upstream API in the future (in the unlikely event that Rails
should ever add a "flashable_error_string" method).

Signed-off-by: Wincent Colaiuta <win@wincent.com>