Make button_to_delete_topic unobtrusive
Continuing in the line of recent commits, make this method participate
in the unobtrusive "button_to_delete_model" pattern.

Note that this one requires a little more work than the previous commit
(which changed the "button_to_delete_comment" method), because of the
way the routes for this nested resource are declared:

In order to keep the method simple, we have to allow a new, shallow
route.

Signed-off-by: Wincent Colaiuta <win@wincent.com>