Make button_to_destroy_page method unobtrusive
Note that we can't just re-use the button_to_destroy_model method here
as we must pass in a special URL due to the way this resource is nested.

We can't employ the shallow route trick that we did for the
button_to_destroy_topic method, because the page model uses permalinks
and so the "to_param" method will not give us an appropriate result;
that is, given two products, each with an "about" page but a different
underlying ID, instead of getting shallow routes like "/pages/1" and
"/pages/2" we will get "/products/foo/page/about" and
"/products/bar/page/about".

Note also that seeing as we are replacing the old AJAX-y button with a
simple form post and no fade-out afterwards, we need to add a
format.html block to handle this case.

Signed-off-by: Wincent Colaiuta <win@wincent.com>