Prefer ===/!== operators over ==/!= in JavaScript
Use of == and != can lead to subtle bugs because these operators cause
implicit type coercion which may lead to unexpected or surprising
results.

While I don't think there actually were any such bugs in the changed
code, it is good practice to use === and !== in general so as to avoid
this kind of surprise.

Signed-off-by: Wincent Colaiuta <win@wincent.com>