First cut at restoring AJAX to unobstrusive model deletion
Still work to be done but the basics are there: note that we annotate
the "button_to" call with info about the DOM ID of the model element
that we wish to fade out once the call has succeeded.

Note that there are bugs: the "js" format isn't detected at all and so
our "html" format code gets executed, which means a flash message is
added which we don't want here. I am thinking of changing this so that I
return some JavaScript (form a template), and changing our request to
actually post to "issues/1.js" rather than "issues/1". This in turn
would make our helper code simpler and the unobtrusive JavaScript
simpler also.

Signed-off-by: Wincent Colaiuta <win@wincent.com>