Be more radical about avoiding HTML layout around non-HTML requests
In commit 5886d0df8 we broadened the checks used to decide when to
render without a layout to include JavaScript requests (previously only
Atom requests were checked for).

It turns out that requests such as:

  https://wincent.com/js/issue/show.js

And coming in with a MIME-type of "*/*" and so were being embedded in an
HTML layout, breaking them.

So to avoid this kind of nonsense in the future, just rendering
everything without a layout, except stuff that is explicitly marked as
having HTML format.

Signed-off-by: Wincent Colaiuta <win@wincent.com>