Inline template_exists? method in JsController
Not sure why this is working on the production server, but not here, but
what I'm seeing is the "render @template" call is failing because deep
in the bowels of ActionController, it is trying to call
"template_exists?", which is a method called from:

  actionpack-3.0.3/lib/abstract_controller/layouts.rb

and which takes two parameters.

However, as we have a method of the same name in the JsController, that
method is getting invoked instead, and is causing an exception because
it expects zero parameters, not two.

As this little helper method is only one line, and it is only used once,
just inline it. The mystery of why I am seeing this locally but not on
the production machine shall remain so, as it is 5 AM and I am in no
mood to start bisecting.

Signed-off-by: Wincent Colaiuta <win@wincent.com>