Temporarily revert 840b85b (layout suppression for non-HTML requests)
Commit 840b85b unfortunately breaks quite a few unit tests, because the
headless browser used in the acceptance specs ends up sending an accept
header with a "*/*" MIME type. This in turn means that "format.html?"
returns false, and no layout is rendered for those specs, breaking any
spec which is looking for stuff in the header, or which depends on
JavaScript stuff included in the head.

So for now, revert, and go back to forcing the layout to "false" in our
JsController. In the longer term, I want to figure out why my requests
to that controller for URLs like "/js/issues/show.js" aren't having the
format correctly set, despite the explicit extension.

Signed-off-by: Wincent Colaiuta <win@wincent.com>