Make #button_to_destroy_model degrade gracefully
By linking to the standard (HTML format) URL and only appending the "js"
extension via JavaScript, we ensure graceful degradation.

When JavaScript is turned off the format.html block will be triggered
(showing a flash and redirecting to an appropriate place) and when it is
turned on we get the normal AJAX behavior (usually a fade out).

Signed-off-by: Wincent Colaiuta <win@wincent.com>