Remove superfluous button_to_destroy_{issue,topic,comment} helpers
All of these are effectively the same now, so we may as well just use
the button_to_destroy_model method directly.

The only remaining special case helper is the button_to_destroy_page
method, which could also be removed if we were prepared to pass in the
custom URL at each call site; at this stage there is only one such call
site, so I'll probably go ahead and do this.

Signed-off-by: Wincent Colaiuta <win@wincent.com>