Use before block instead of helper in commentable specs
The comment here about the before block not working because it caused
items to be left behind in the database between specs was false; the
real reason was that some of the specs wanted to start with the
assumption tha the database was empty.

Fix that, replacing the helper method with a standard before block, and
then using the:

  expect do
    something!
  end.to change { something }.by(1)

pattern. Note that in doing this clean-up I also moved the repeated
call to #reload up into the #add_comment method.

Signed-off-by: Wincent Colaiuta <win@wincent.com>