Ruby 1.9: Pathname needs an explicit to_s
I've recently experimented with moving to Ruby 1.9 but there were some
issues with it which mean that I won't bother pursuing it much further
right now.

Nevertheless this is one small, harmless change that I discovered which
will eventually be necessary for 1.9, and which does no harm under 1.8,
so may as well get it into the tree already.

Signed-off-by: Wincent Colaiuta <win@wincent.com>