Add #redirected_from helper for custom admin logic
For articles that we arrive at as a result of a redirection, we want to
show not only the title, but for admins we want to show a separate edit
link as well.

Previously we made the title itself the edit link, but that was
confusing and looked a little like a bug. By keeping it separate it
becomes clear that the intention is not only to show the title, but
offer admins an opportunity of fixing bad redirections should they crop
up.

See:

  https://wincent.com/issues/1803

Signed-off-by: Wincent Colaiuta <win@wincent.com>