Be as relaxed as possible in specifying versions in Gemfile
Trust the Gemfile.lock to keep us pinned to versions which actually
work; "bundle update" should be fine, and our test suite should uncover
any compatibility problems when we update. Only peg a Gem to a
particular version in the Gemfile if we have a known bug that is forcing
us to stay at that version.

Signed-off-by: Wincent Colaiuta <win@wincent.com>