Refactor authentication logic
Seeking to clean up some code smell here. The main change is renaming
the #set_current_user= method to #set_current_user, because the
duplication of the verb "set" and the setter "=" was redundant.

We also rewrite the comments to make it clear what the difference is
between #set_current_user (an explicit log-in action, taken by the user,
such as submitting a login form or clicking a logout link, and called
from the controller) and the #current_user= method (an automatic
process, generally performed from a before filter as part of the request
cycle, such as setting the current user based on a cookie or similar).

Signed-off-by: Wincent Colaiuta <win@wincent.com>