Fix new builder view spec failures under Rails 3.1
It seems that specifying "action.atom" is no longer enough to get
Rails/RSpec to render the right template in view tests; you have to pass
the entire filename (with the ".builder" extension) or it will just go
ahead and render the HTML version of the template.

While being explicit about the template is not a bad thing, still seems
like a bug that it would render "action.html.haml" rather than
"action.atom.builder" when asked for "action.atom". At the very least,
one would hope it would throw an error rather than proceeding to render
the wrong template.

Signed-off-by: Wincent Colaiuta <win@wincent.com>